All ratings in ints now. avg to aPCG has revealed stdDev error.

+++ IPCG/src/Model/Level.java
@@ -17,8 +17,8 @@ public class Level extends GameCollection<Pattern> {
 	public void tweak() { }
 
 	@Override
-	public double rate() { 
-		double total = 0;
+	public int rate() { 
+		int total = 0;
 		for (Pattern p : elements) {
 			total += p.rate();
 		}
